Better user support in README and Dockerfile
In the README I included instructions for the different platforms
encourage usage in Mac, Linux and Windows with docker.

Modified the Dockerfile to put all the apt-get calls on one call with
alphabetical order to follow best practices. Also moved the db4 download
to later in the process because it invalidates the docker build cache
once we use it.